D3ASIM-1024: Adapted to PR comments.
