+++ integration_tests/steps/jira_issues_tests.py
@@ -17,8 +17,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 from behave import then
 from math import isclose
-from d3a.d3a_core.sim_results.export_unmatched_loads import ExportUnmatchedLoads
-from d3a.d3a_core.util import get_number_of_unmatched_loads
+from d3a.d3a_core.sim_results.export_unmatched_loads import ExportUnmatchedLoads, \
+    get_number_of_unmatched_loads
 
 
 def get_areas_from_2_house_grid(context):
@@ -123,7 +123,7 @@ def check_matching_trades(context):
 
 @then('there should be no unmatched loads')
 def no_unmatched_loads(context):
-    unmatched, unmatched_redis = ExportUnmatchedLoads(context.simulation.area).__call__()
+    unmatched, unmatched_redis = ExportUnmatchedLoads(context.simulation.area)
     assert get_number_of_unmatched_loads(unmatched) == 0
